VPLAY-10805 class sep and main_aamp dep untangling
Reason for Change: teh debt
