+++ iso639map.h
@@ -24,10 +24,7 @@
  * @file iso639map.h
  * @brief ISO639 is a standard with representation of names for languages 
  */
-
-
-#include <stdio.h>
-#include "main_aamp.h"
+#include "LangCodePreference.h"
 
 void iso639map_NormalizeLanguageCode( char lang[], LangCodePreference langCodePreference );
