+++ test/utests/fakes/FakePrivateInstanceAAMP.cpp
@@ -148,6 +148,16 @@ PrivateInstanceAAMP::~PrivateInstanceAAMP()
 {
 }
 
+double PrivateInstanceAAMP::RecalculatePTS(AampMediaType mediaType, const void *ptr, size_t len)
+{
+    double pts = 0.0;
+    if (g_mockPrivateInstanceAAMP != nullptr)
+    {
+        pts = g_mockPrivateInstanceAAMP->RecalculatePTS(mediaType, ptr, len);
+    }
+    return pts;
+}
+
 size_t PrivateInstanceAAMP::HandleSSLWriteCallback ( char *ptr, size_t size, size_t nmemb, void* userdata )
 {
 	return 0;
