+++ test/utests/mocks/MockPrivateInstanceAAMP.h
@@ -26,6 +26,8 @@
 class MockPrivateInstanceAAMP
 {
 public:
+    MOCK_METHOD(double, RecalculatePTS, (AampMediaType mediaType, const void *ptr, size_t len));
+    
 	MOCK_METHOD(void, Stop, (bool sendStateChangeEvent));
 
 	MOCK_METHOD(void, StartPausePositionMonitoring, (long long pausePositionMilliseconds));
