+++ test/utests/tests/AampTSBSessionManager/FunctionalTests.cpp
@@ -176,6 +176,8 @@ TEST_F(FunctionalTests, TSBWriteTests)
 
 	EXPECT_CALL(*g_mockTSBStore, Write(UNIQUE_INIT_URL, TEST_DATA, strlen(TEST_DATA))).WillOnce(Return(TSB::Status::OK));
 	EXPECT_CALL(*g_mockPrivateInstanceAAMP, GetVidTimeScale()).WillRepeatedly(Return(1));
+    //EXPECT_CALL(*g_mockPrivateInstanceAAMP, RecalculatePTS(eMEDIATYPE_INIT_VIDEO,_,_)).Times(1).WillOnce(Return(0.0));
+    //EXPECT_CALL(*g_mockPrivateInstanceAAMP, RecalculatePTS(eMEDIATYPE_VIDEO,_,_)).Times(2).WillRepeatedly(Return(1234));
 	mAampTSBSessionManager->EnqueueWrite(INIT_URL, cachedFragment, TEST_PERIOD_ID);
 	std::this_thread::sleep_for(std::chrono::milliseconds(25));
 
@@ -254,6 +256,8 @@ TEST_F(FunctionalTests, Cullsegments)
 	EXPECT_CALL(*g_mockTSBStore, Write(_,_,_)).WillRepeatedly(Return(TSB::Status::OK));
 	EXPECT_CALL(*g_mockPrivateInstanceAAMP, GetVidTimeScale()).WillRepeatedly(Return(1));
 	EXPECT_CALL(*g_mockPrivateInstanceAAMP, GetAudTimeScale()).WillRepeatedly(Return(1));
+    //EXPECT_CALL(*g_mockPrivateInstanceAAMP, RecalculatePTS(eMEDIATYPE_INIT_VIDEO,_,_)).Times(1).WillOnce(Return(0.0));
+    //EXPECT_CALL(*g_mockPrivateInstanceAAMP, RecalculatePTS(eMEDIATYPE_VIDEO,_,_)).Times(2).WillRepeatedly(Return(4567));
 
 	const std::string initUrl = std::string(TEST_BASE_URL) + std::string("init.mp4");
 	cachedFragment->type = eMEDIATYPE_INIT_VIDEO;
@@ -330,8 +334,8 @@ TEST_F(FunctionalTests, TSBReadTests)
 
 	EXPECT_CALL(*g_mockTSBStore, Write(_,_,_)).WillRepeatedly(Return(TSB::Status::OK));
 	EXPECT_CALL(*g_mockPrivateInstanceAAMP, GetVidTimeScale()).WillRepeatedly(Return(1));
-	EXPECT_CALL(*g_mockAampUtils, RecalculatePTS(eMEDIATYPE_INIT_VIDEO,_,_,_)).Times(1).WillOnce(Return(0.0));
-	EXPECT_CALL(*g_mockAampUtils, RecalculatePTS(eMEDIATYPE_VIDEO,_,_,_)).Times(2).WillRepeatedly(Return(FRAG_FIRST_PTS));
+	EXPECT_CALL(*g_mockPrivateInstanceAAMP, RecalculatePTS(eMEDIATYPE_INIT_VIDEO,_,_)).Times(1).WillOnce(Return(0.0));
+	EXPECT_CALL(*g_mockPrivateInstanceAAMP, RecalculatePTS(eMEDIATYPE_VIDEO,_,_)).Times(2).WillRepeatedly(Return(FRAG_FIRST_PTS));
 
 	const std::string initUrl = std::string(TEST_BASE_URL) + std::string("init.mp4");
 	const std::string videoUrl = std::string(TEST_BASE_URL) + std::string("video.mp4");
